Use SlicerOpenLIFU's python requirements
Issue #2
Use the python requirements file specified in the SlicerOpenLIFU
extension rather than specifying it separately here.
This way there is one source of truth as to which version of the
openlifu package to use -- it's the one specified by the version of
the SlicerOpenLIFU extension that we bundle into the app.